working on bugs

+++ BlackJack/Models/Player.cs
@@ -9,7 +9,7 @@ namespace BlackJack
     class Player
     {
         public string Name { get; set; }
-        public int Cash { get; set; }
+        public double Cash { get; set; }
         public List<Card> Hand { get; set; }
         public int HandValue { get; set; }
         public int Wager { get; set; }
